Sprint2

+++ berza-servis/src/main/java/rs/edu/raf/opcija/servis/util/OptionYahooApiMap.java
@@ -56,6 +56,6 @@ public class OptionYahooApiMap {
 
     private Long volume;
 
-    private Long openInterest;
+    private Double openInterest;
 
 }
